# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Social media platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /users/{userId}/posts:
    get:
      operationId: listUserPosts
      tags: [users]
      summary: Get posts by user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
  /users/{userId}/followers:
    get:
      operationId: getFollowers
      tags: [follows]
      summary: Get followers of a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/{userId}/following:
    get:
      operationId: getFollowing
      tags: [follows]
      summary: Get users followed by a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Following list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/me/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, username, display_name, created_at, followers_count, following_count, posts_count]
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        bio:
          type: ["string", "null"]
        avatar_url:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        created_at:
          type: string
        followers_count:
          type: integer
        following_count:
          type: integer
        posts_count:
          type: integer
        is_following:
          type: boolean
    UserSummary:
      type: object
      required: [id, username, display_name]
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        is_following:
          type: boolean
        followers_count:
          type: integer
    ProfileUpdate:
      type: object
      properties:
        display_name:
          type: string
        bio:
          type: string
        avatar_url:
          type: string
        website:
          type: string
    Post:
      type: object
      required: [id, content, author, created_at, likes_count, comments_count]
      properties:
        id:
          type: string
        content:
          type: string
        image_url:
          type: ["string", "null"]
        author:
          $ref: "#/components/schemas/UserSummary"
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
        likes_count:
          type: integer
        comments_count:
          type: integer
        is_liked:
          type: boolean
...
    PostsPage:
      type: object
      required: [posts, has_more]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        next_cursor:
          type: ["string", "null"]
        has_more:
          type: boolean
...