# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social media platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  display_name?: string;
  bio?: string;
  avatar_url?: string;
  website?: string;
}