# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social media platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: string;
  username: string;
  display_name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatar_url?: string | null;
  /** @nullable */
  website?: string | null;
  created_at: string;
  followers_count: number;
  following_count: number;
  posts_count: number;
  is_following?: boolean;
}